% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAMultipleAlignment_to_df.R
\name{AAMultipleAlignment_to_df}
\alias{AAMultipleAlignment_to_df}
\title{AAMultipleAlignment_to_df (TESTING)}
\usage{
AAMultipleAlignment_to_df(AAS_align)
}
\arguments{
\item{AAS_align}{Multiple alignment object}
}
\value{
DF a data.frame containing the sequences. Rows represent individual sequences, columns the sequence "items" (i.e. one base per column).
}
\description{
This function is designed to convert Multiple Alignments of the XStringSet class into a data.frame
}
