% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ab1Plot.R
\name{Ab1Plot}
\alias{Ab1Plot}
\title{OverlapAnalysis}
\usage{
Ab1Plot(file = "15.4I.ab1", range, ratio = NULL)
}
\arguments{
\item{file}{path to Ab1 file}

\item{range}{limits of range to print ab1 plot}
}
\description{
Make Ab1 plot
}
