% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addROItoGraph.R
\name{addROItoGraph}
\alias{addROItoGraph}
\title{addROItoGraph (TESTING)}
\usage{
addROItoGraph(plot, ROIs, ROIfill = "blue")
}
\arguments{
\item{plot}{ggplot to append to}

\item{ROIs}{Object of type GRranges with ROI(s)}

\item{ROIfill}{ROI fill colour}
}
\description{
Adds a ROI box from IRanges to a plot
}
