% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAMultipleAlignment_to_df.R
\name{DNAMultipleAlignment_to_df}
\alias{DNAMultipleAlignment_to_df}
\title{DNAMultipleAlignment_to_df (TESTING)}
\usage{
DNAMultipleAlignment_to_df(DNAS_align)
}
\arguments{
\item{DNAS_align}{Multiple alignment object}
}
\value{
DF a data.frame containing the sequences. Rows represent individual sequences, columns the sequence "items" (i.e. one base per column).
}
\description{
This function is designed to convert Multiple Alignments of the XStringSet class into a data.frame
}
