% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makegenotypeidentifiers.R
\name{makegenotypeidentifiers}
\alias{makegenotypeidentifiers}
\title{A function to extractg the genotpe based on ab1 files}
\usage{
makegenotypeidentifiers(litter_list, size_list, alpha = TRUE)
}
\arguments{
\item{litter_list}{A list of litter names}

\item{size_list}{Corresponding list with number of pubs}

\item{alpha}{Make alphabetic instead of numeric idenfiers}
}
\description{
This function retrieves ab1 files either from a link or from a directory and compares the sequence to .
}
