% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeseq.R
\name{sangergenotype}
\alias{sangergenotype}
\title{A function to extractg the genotpe based on ab1 files}
\usage{
sangergenotype(dir = choose.dir(), link = "", wtseq = "",
  mutseq = "", revcomp = TRUE, cutoff = 0.33, separator = "_")
}
\arguments{
\item{dir}{Source directory for .ab1 files tp analyze or, if link ist not blank, directory where to download and unpack the archive containing .ab1 files to. In that case, a new subdirectory named by the current date will be created. If left blank, a file open dialog will appear.}

\item{link}{Optional. A url to a archive containing the ab1 files to analyze.}

\item{wtseq}{Sequence for wild-type}

\item{mutseq}{Sequence for mutant}

\item{revcomp}{make reverse complementary of wtseq and mutseq}

\item{cutoff}{minimum relative amplitude of secondary peak}

\item{separator}{A separator char or string by which the ab1 filenames should be split and analyzed in order to identify unique strings (e.g. mouse identifiers)}
}
\value{
Data Frame with list of mice and respective genotypes
}
\description{
This function retrieves ab1 files either from a link or from a directory and compares the sequence to .
}
\examples{
sangergenotype(dir="D:\\\\", link="http://www..zip",wtseq="ACTG",mutseq="ACTG", revcomp = TRUE, cutoff = 0.2)
}
