% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vizualize_MultAlignment.R
\name{Vizualize_MultAlignment}
\alias{Vizualize_MultAlignment}
\title{Export Function}
\usage{
Vizualize_MultAlignment(AAS, restrict_to = FALSE, restrict_by,
  sortbyIdentity = TRUE, DistanceMatrix = NULL, ROI = NULL)
}
\arguments{
\item{AAS}{AAStringSet containing sequences to align}

\item{restrict_to}{FALSE=prints full alignment, "Sequence": only plot range, where sequence given in @param restrict_by is aligned. "Range", only plot range given in restrict_by}

\item{sortbyIdentity}{if restrict_to="Sequence", order by similarity to the Sequence named in restrict_by}

\item{DistanceMatrix}{if restrict_to="Sequence", optional 2x lengh of Sequene named in restrict_by matrix arranging the x position of the reads. useful for plotting alongside ab1 grahps}

\item{ROI}{Range of pos values. If given, marks region on plot in the defined range.}
}
\value{
ggplot2 showing aligned sequences
}
\description{
Converts Multiple Alignments object to data.frame
}
